# app-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это сервер для книги рецептов (лабораторная работа).",
        "title": "Recipe Book API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/dishes": {
            "get": {
                "description": "Возвращает блюда с фильтрацией по категории, флагам и имени.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Получить список блюд",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Категория блюда",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Флаги блюда",
                        "name": "flags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Подстрока для поиска по имени",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.dishesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новое блюдо с составом продуктов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Создать блюдо",
                "parameters": [
                    {
                        "description": "Данные блюда",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createDishRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1.dishResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/dishes/{id}": {
            "put": {
                "description": "Частично обновляет данные блюда по UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Обновить блюдо",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID блюда",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для обновления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.updateDishRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.messageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет блюдо по UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Удалить блюдо",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID блюда",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.messageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/dishes/{id}/products": {
            "get": {
                "description": "Получить продукты, которые используются для приготовления блюда с UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Получить продукты входящие в состав блюда.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID блюда",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.messageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Dish": {
            "type": "object",
            "properties": {
                "calories": {
                    "type": "number"
                },
                "carbs": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/domain.DishCategory"
                },
                "created_at": {
                    "type": "string"
                },
                "fats": {
                    "type": "number"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "proteins": {
                    "type": "number"
                },
                "serving_size": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.DishCategory": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7
            ],
            "x-enum-varnames": [
                "DishCategoryDessert",
                "DishCategoryMainCourse",
                "DishCategoryAppetizer",
                "DishCategorySalad",
                "DishCategorySoup",
                "DishCategoryDrink",
                "DishCategoryStarter",
                "DishCategoryUnknown"
            ]
        },
...
        "usecase.DishProductDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
...
        "v1.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.createDishRequest": {
            "type": "object",
            "required": [
                "calories",
                "carbs",
                "category",
                "fats",
                "name",
                "products",
                "proteins",
                "serving_size"
            ],
            "properties": {
                "calories": {
                    "type": "number"
                },
                "carbs": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/domain.DishCategory"
                },
                "fats": {
                    "type": "number"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.DishProductDTO"
                    }
                },
                "proteins": {
                    "type": "number"
                },
                "serving_size": {
                    "type": "number"
                }
            }
        },
...
        "v1.dishData": {
            "type": "object",
            "properties": {
                "dish": {
                    "$ref": "#/definitions/domain.Dish"
                }
            }
        },
        "v1.dishResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.dishData"
                },
                "error": {
                    "$ref": "#/definitions/v1.apiError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "v1.dishesData": {
            "type": "object",
            "properties": {
                "dishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Dish"
                    }
                }
            }
        },
        "v1.dishesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.dishesData"
                },
                "error": {
                    "$ref": "#/definitions/v1.apiError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "v1.errorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/v1.apiError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "v1.messageData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.messageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.messageData"
                },
                "error": {
                    "$ref": "#/definitions/v1.apiError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "v1.updateDishRequest": {
            "type": "object",
            "properties": {
                "calories": {
                    "type": "number"
                },
                "carbs": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/domain.DishCategory"
                },
                "fats": {
                    "type": "number"
                },
                "flags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.DishProductDTO"
                    }
                },
                "proteins": {
                    "type": "number"
                },
                "serving_size": {
                    "type": "number"
                }
            }
        }
    }
}